# Reduce indirection and memory overhead for slot maps.