# Reducing the property related memory overhead of JavaScript objects.

## Current slot map model

Currently JavaScript Objects in RHino store their properties in slot map, which is held by a slot map container. The map itself then either holds an array, or some other collection structure, which actually holds the slots. So the structure looks something like this:

```mermaid
graph LR
Object["ScriptableObject"]
Container["SlotMapContainer"]
Map["SlotMap"]
Array["Slot[]"]
Slot1["Slot"]
Slot2["Slot"]
Slot3["Slot"]
Object1["ScriptableObject"]
Container1["SlotMapContainer"]
Map1["SlotMap"]
Array1["Slot[]"]
Slot4["Slot"]
Slot5["Slot"]
Slot6["Slot"]
Object --> Container --> Map --> Array
Array --> Slot1
Array --> Slot2
Array --> Slot3
Object1 --> Container1 --> Map1 --> Array1
Array1 --> Slot4
Array1 --> Slot5
Array1 --> Slot6
```

And in any environment which relies on thread safe slot map handling there is an additional stamped lock object like this:

```mermaid
graph LR
Object["ScriptableObject"]
Container["SlotMapContainer"]
Lock["StampedLock"]
Map["SlotMap"]
Array["Slot[]"]
Slot1["Slot"]
Slot2["Slot"]
Slot3["Slot"]
Object --> Container --> Map --> Array
Container --> Lock
Array --> Slot1
Array --> Slot2
Array --> Slot3
Object1["ScriptableObject"]
Container1["SlotMapContainer"]
Lock1["StampedLock"]
Map1["SlotMap"]
Array1["Slot[]"]
Slot4["Slot"]
Slot5["Slot"]
Slot6["Slot"]
Object1 --> Container1 --> Map1 --> Array1
Container1 --> Lock1
Array1 --> Slot4
Array1 --> Slot5
Array1 --> Slot6
```

There is one optimisation here, in that the arrays for slots in an `EmbeddedSlotMap` are not created until slots are stored in the map, but all the other objects are created eagerly. If I look at a heap dump from one of our production systems and analyse the number of properties stored in slot maps then we see the following:

| No. of props | Count   | Cummulative count | Cummulative % |
|--------------|---------|-------------------|---------------|
| 0            | 113,538 | 113,538           | 76.6428827    |
| 1            | 28,013  | 141,551           | 95.5528254    |
| 2            | 1,243   | 142,794           | 96.3919022    |
| 3            | 1,965   | 144,759           | 97.7183591    |
| 4            | 426     | 145,185           | 98.0059269    |

The remaining objects are distributed over a fairly wide range.

The prevalence of zero size slot maps may be surprising, but is due to the large number of common objects that are implemented in Rhino itself as `IdScriptableObject`s with built in properties not held as real slots (functions, strings, arrays, wrapped native Java methods, etc.). This means we pay a significant memory overhead on the 75% of objects that will never have properties, and we aren't optimising for the remaining commons case of a single property. This type of property map size distribution appears consistent with my experience on TruffleRuby and other dynamic language implementations, so I would expect roughly similar distributions for other users.

In our particular case we have a small number of objects shared between threads, and so enable thread safety, so we pay for a lock with every one of these objects.

The memory overheads we incur are as follows (on a 64bit JVM with compact oops

| Object             | Size                       |
|--------------------|----------------------------|
| `SlotMapContainer` | 24                         |
| `StampedLocked`    | 48                         |
| `EmbeddedSlotMap`  | 32                         |
| `Slot[]`           | 16 + 4 * (size rounded up) |

Combined with the data on object sizes about this gives us total sizes as follows

| No. of props | Object size | Thread safe object size | Count  | Size * Count | Thread safe size * Count |
|--------------|-------------|-------------------------|--------|--------------|--------------------------|
|            0 |          56 |                     104 | 113538 |      6358128 |                 11807952 |
|            1 |          88 |                     136 |  28013 |      2465144 |                  3809768 |
|            2 |          88 |                     136 |   1243 |       109384 |                   169048 |
|            3 |          88 |                     136 |   1965 |       172920 |                   267240 |
|            4 |          88 |                     136 |    426 |        37488 |                    57936 |

## Reducing indirection

Ideally we would want the common cases to incur as little overhead as possible, and potential store a small number of properties within the `ScriptableObject` itself if that significantly reduced overhead. Short of that we can optimise for the common cases and change the existing slot map promotion chain of:

```mermaid
graph LR
EmbededdedSlotMap --> HashSlotMap
```

to something more like:

```mermaid
graph LR
EmptySlotMap --> SingleEntrySlotMap --> EmbededdedSlotMap --> HashSlotMap
```

I do not propose to embed slots in the `ScriptableObject`s themselves at this time, but instead exploit the fact that we can use a singular immutable empty slot map; refactor out the slot map container for single threaded use; introduce a single entry slot map to reduce overhead and indirection in the remaining common case; and finally remove the use of slot map containers in the multi-threaded case.

## PRs

I've split this into a number of stacked PRs detailed below
* #1782 (This PR)
* #1783
* #1784
* #1785

### Introduce a singleton `EMPTY_SLOT_MAP` (#1782)

This PR.

Introduces a singleton `EMPTY_SLOT_MAP` reducing the common empty case to
```mermaid
graph LR
Object["ScriptableObject"]
Container["SlotMapContainer"]
Map["EMPTY_SLOT_MAP"]
Object1["ScriptableObject"]
Container1["SlotMapContainer"]
Object --> Container --> Map
Object1 --> Container1 --> Map
```

This saves 32 bytes per object with an empty property map.

### Start to factor out slot map container (#1783)

Introduces the concept of a `SlotMapOwner` and refactors `ScritableObject` and `SlotMapContainer` to inherit from `SlotMapOwner`. The slot maps themselves are made responsible for promotions themselves via the `SlotMapOwner`. As well as reducing the amount of indirection needed to access the slots it also means we can avoid promotion checks except when we know the collection might overflow some internal limit (for example, when the array in `EmbeddedSlotMap` would be expanded). This changes the empty map case to:

```mermaid
graph LR
Object["ScriptableObject"]
Map["EMPTY_SLOT_MAP"]
Object1["ScriptableObject"]
Object --> Map
Object1 --> Map
```

and the occupied map case to

```mermaid
graph LR
Object["ScriptableObject"]
Map["SlotMap"]
Array["Slot[]"]
Slot1["Slot"]
Slot2["Slot"]
Slot3["Slot"]
Object1["ScriptableObject"]
Map1["SlotMap"]
Array1["Slot[]"]
Slot4["Slot"]
Slot5["Slot"]
Slot6["Slot"]
Object --> Map --> Array
Array --> Slot1
Array --> Slot2
Array --> Slot3
Object1 --> Map1 --> Array1
Array1 --> Slot4
Array1 --> Slot5
Array1 --> Slot6
```

### Single-entry slot maps (#1784)

Introduces an immutable single-entry slot map. It is immutable and *always* promotes to `EmbeddedSlotMap` to make its future use in multi-threaded cases easier, but also because it seems like a reasonable assumption that any single-entry slot map that is mutated is likely to continue being mutated.

This changes the single slot case to:

```mermaid
graph LR
Object["ScriptableObject"]
Map["SingleEntrySlotMap"]
Slot1["Slot"]
Object1["ScriptableObject"]
Map1["SingleEntrySlotMap"]
Slot2["Slot"]
Object --> Map --> Slot1
Object1 --> Map1 --> Slot2
```

### Introduce thread safe versions of slot maps (#1785)

Introduces an API on slot map owner to do a compare and swap of the current slot map to enable thread safe promotion of maps, and introduces thread safe versions of the slot map types which handle locking where necessary. We avoid the use of locks on empty and single entry maps as they are immutable, and introduce APIS and share locks between further promotions to avoid the overhead of an additional container.

## Benchmark comparison

Benchmark scores are roughly neutral, some are gains, some gain performance on the two full runs I performed, and some lost performance, but they all show a degree of instability on repeated runs.

| Benchmark                                                   | (interpreted) | Mode | Cnt |  Old Score | Old Error |  New Score |  New Error | Units |      % Gain |    % Error |
|-------------------------------------------------------------|---------------|------|-----|------------|-----------|------------|------------|-------|-------------|------------|
| BuiltinBenchmark.annotatedClassMethods                      | false         | avgt |   5 |     21.521 |     0.215 |     25.207 |      0.359 | us/op |  -17.127457 |  1.4242076 |
| BuiltinBenchmark.annotatedClassMethods                      | true          | avgt |   5 |    146.907 |    18.851 |    125.198 |      1.876 | us/op |   14.777376 |  1.4984265 |
| BuiltinBenchmark.dumbLambdaClassMethods                     | false         | avgt |   5 |     18.682 |     0.470 |     18.261 |      0.202 | us/op |   2.2535060 |  1.1061826 |
| BuiltinBenchmark.dumbLambdaClassMethods                     | true          | avgt |   5 |    149.805 |     4.646 |    123.110 |      1.675 | us/op |   17.819832 |  1.3605718 |
| BuiltinBenchmark.idClassMethods                             | false         | avgt |   5 |     25.316 |     1.140 |     22.394 |      2.280 | us/op |   11.542108 |  10.181299 |
| BuiltinBenchmark.idClassMethods                             | true          | avgt |   5 |    141.918 |     5.423 |    122.982 |      1.360 | us/op |   13.342916 |  1.1058529 |
| MathBenchmark.addConstantFloats                             | false         | avgt |   5 |      8.819 |     0.248 |      8.543 |      0.111 | ns/op |   3.1296065 |  1.2993094 |
| MathBenchmark.addConstantFloats                             | true          | avgt |   5 |    192.417 |     7.392 |    176.734 |      6.012 | ns/op |   8.1505272 |  3.4017224 |
| MathBenchmark.addConstantInts                               | false         | avgt |   5 |      8.258 |     0.146 |      8.469 |      0.126 | ns/op |  -2.5550981 |  1.4877790 |
| MathBenchmark.addConstantInts                               | true          | avgt |   5 |    182.400 |     0.414 |    186.263 |      1.745 | ns/op |  -2.1178728 | 0.93684736 |
| MathBenchmark.addIntAndConstant                             | false         | avgt |   5 |     11.183 |     0.680 |     11.310 |      0.142 | ns/op |  -1.1356523 |  1.2555261 |
| MathBenchmark.addIntAndConstant                             | true          | avgt |   5 |    274.836 |     3.912 |    276.018 |      8.103 | ns/op | -0.43007466 |  2.9356781 |
| MathBenchmark.addMixedStrings                               | false         | avgt |   5 |     36.016 |     0.971 |     36.450 |      0.619 | ns/op |  -1.2050200 |  1.6982167 |
| MathBenchmark.addMixedStrings                               | true          | avgt |   5 |    278.977 |     3.535 |    269.311 |      3.058 | ns/op |   3.4648018 |  1.1354902 |
| MathBenchmark.addStringsInLoop                              | false         | avgt |   5 |     84.544 |     7.452 |     84.872 |      2.305 | ns/op | -0.38796366 |  2.7158545 |
| MathBenchmark.addStringsInLoop                              | true          | avgt |   5 |    843.004 |    13.554 |    909.288 |     46.029 | ns/op |  -7.8628334 |  5.0620925 |
| MathBenchmark.addTwoFloats                                  | false         | avgt |   5 |     24.307 |    49.848 |     14.623 |      5.111 | ns/op |   39.840375 |  34.951788 |
| MathBenchmark.addTwoFloats                                  | true          | avgt |   5 |    355.543 |     7.334 |    370.175 |     24.142 | ns/op |  -4.1153953 |  6.5217802 |
| MathBenchmark.addTwoInts                                    | false         | avgt |   5 |     14.168 |     0.225 |     15.857 |      0.199 | ns/op |  -11.921231 |  1.2549663 |
| MathBenchmark.addTwoInts                                    | true          | avgt |   5 |    367.576 |    11.698 |    349.172 |      3.876 | ns/op |   5.0068557 |  1.1100546 |
| MathBenchmark.bitwiseAnd                                    | false         | avgt |   5 |     14.792 |     1.144 |     14.448 |      0.263 | ns/op |   2.3255814 |  1.8203212 |
| MathBenchmark.bitwiseAnd                                    | true          | avgt |   5 |    359.100 |     3.495 |    353.973 |      2.887 | ns/op |   1.4277360 | 0.81559893 |
| MathBenchmark.bitwiseLsh                                    | false         | avgt |   5 |     14.166 |     0.181 |     15.366 |      0.185 | ns/op |  -8.4709869 |  1.2039568 |
| MathBenchmark.bitwiseLsh                                    | true          | avgt |   5 |    340.105 |     9.303 |    307.955 |     17.735 | ns/op |   9.4529631 |  5.7589583 |
| MathBenchmark.bitwiseOr                                     | false         | avgt |   5 |     14.521 |     0.117 |     15.786 |      4.163 | ns/op |  -8.7115212 |  26.371468 |
| MathBenchmark.bitwiseOr                                     | true          | avgt |   5 |    365.443 |     3.158 |    349.798 |      1.803 | ns/op |   4.2811054 | 0.51544034 |
| MathBenchmark.bitwiseRsh                                    | false         | avgt |   5 |     13.753 |     0.148 |     14.356 |      0.162 | ns/op |  -4.3844979 |  1.1284480 |
| MathBenchmark.bitwiseRsh                                    | true          | avgt |   5 |    328.926 |     3.579 |    311.810 |      4.052 | ns/op |   5.2036020 |  1.2995093 |
| MathBenchmark.bitwiseSignedRsh                              | false         | avgt |   5 |     13.575 |     0.261 |     14.134 |      0.158 | ns/op |  -4.1178637 |  1.1178718 |
| MathBenchmark.bitwiseSignedRsh                              | true          | avgt |   5 |    360.686 |     6.364 |    347.190 |      2.527 | ns/op |   3.7417588 | 0.72784354 |
| MathBenchmark.subtractFloats                                | false         | avgt |   5 |     11.270 |     0.119 |     11.309 |      0.114 | ns/op | -0.34605146 |  1.0080467 |
| MathBenchmark.subtractFloats                                | true          | avgt |   5 |    280.508 |     9.309 |    271.140 |      2.755 | ns/op |   3.3396552 |  1.0160803 |
| MathBenchmark.subtractInts                                  | false         | avgt |   5 |     11.780 |     0.558 |     11.865 |      0.131 | ns/op | -0.72156197 |  1.1040877 |
| MathBenchmark.subtractInts                                  | true          | avgt |   5 |    274.109 |    16.658 |    272.634 |      2.526 | ns/op |  0.53810710 | 0.92651687 |
| MathBenchmark.subtractTwoFloats                             | false         | avgt |   5 |     14.159 |     0.572 |     14.389 |      0.107 | ns/op |  -1.6244085 | 0.74362360 |
| MathBenchmark.subtractTwoFloats                             | true          | avgt |   5 |    350.104 |     9.883 |    346.064 |      4.471 | ns/op |   1.1539428 |  1.2919576 |
| ObjectBenchmark.accessFields                                | false         | avgt |   5 |      0.062 |     0.001 |      0.073 |      0.005 | us/op |  -17.741935 |  6.8493151 |
| ObjectBenchmark.accessFields                                | true          | avgt |   5 |      0.507 |     0.020 |      0.557 |      0.005 | us/op |  -9.8619329 | 0.89766607 |
| ObjectBenchmark.createFields                                | false         | avgt |   5 |     46.803 |     0.878 |     47.669 |      0.320 | us/op |  -1.8503087 | 0.67129581 |
| ObjectBenchmark.createFields                                | true          | avgt |   5 |    318.915 |     4.921 |    348.613 |      5.578 | us/op |  -9.3121992 |  1.6000551 |
| ObjectBenchmark.deleteFields                                | false         | avgt |   5 |      0.066 |     0.001 |      0.073 |      0.001 | us/op |  -10.606061 |  1.3698630 |
| ObjectBenchmark.deleteFields                                | true          | avgt |   5 |      0.475 |     0.005 |      0.523 |      0.014 | us/op |  -10.105263 |  2.6768642 |
| ObjectBenchmark.iterateFields                               | false         | avgt |   5 |     82.580 |     1.017 |     83.256 |      4.002 | us/op | -0.81860015 |  4.8068608 |
| ObjectBenchmark.iterateFields                               | true          | avgt |   5 |    227.787 |    28.930 |    230.405 |      2.585 | us/op |  -1.1493193 |  1.1219375 |
| ObjectBenchmark.ownKeysFields                               | false         | avgt |   5 |     79.724 |     0.650 |     78.692 |      0.928 | us/op |   1.2944659 |  1.1792812 |
| ObjectBenchmark.ownKeysFields                               | true          | avgt |   5 |     81.579 |     7.778 |     79.953 |      1.003 | us/op |   1.9931600 |  1.2544870 |
| PropertyBenchmark.addTwoProperties                          | false         | avgt |   5 |      8.999 |     0.129 |      8.763 |      0.137 | ns/op |   2.6225136 |  1.5633915 |
| PropertyBenchmark.addTwoProperties                          | true          | avgt |   5 |    101.038 |     1.496 |    106.104 |      1.081 | ns/op |  -5.0139551 |  1.0188117 |
| PropertyBenchmark.createObject                              | false         | avgt |   5 |    169.871 |     2.012 |     96.735 |      1.937 | ns/op |   43.053847 |  2.0023776 |
| PropertyBenchmark.createObject                              | true          | avgt |   5 |    240.923 |     3.740 |    206.758 |      1.446 | ns/op |   14.180879 | 0.69936834 |
| PropertyBenchmark.createObjectFieldByField                  | false         | avgt |   5 |    169.260 |     4.393 |     85.181 |      1.198 | ns/op |   49.674465 |  1.4064169 |
| PropertyBenchmark.createObjectFieldByField                  | true          | avgt |   5 |    324.056 |     5.682 |    377.576 |      6.570 | ns/op |  -16.515664 |  1.7400470 |
| PropertyBenchmark.getOneProperty                            | false         | avgt |   5 |      7.139 |     0.073 |      7.306 |      0.096 | ns/op |  -2.3392632 |  1.3139885 |
| PropertyBenchmark.getOneProperty                            | true          | avgt |   5 |     34.944 |     0.642 |     35.038 |      0.605 | ns/op | -0.26900183 |  1.7266967 |
| SlotMapBenchmark.embeddedInsert1Key                         | N/A           | avgt |   5 |      1.047 |     0.020 |      1.061 |      0.013 | ns/op |  -1.3371538 |  1.2252592 |
| SlotMapBenchmark.embeddedQueryKey100Entries                 | N/A           | avgt |   5 |      0.805 |     0.057 |      0.804 |      0.008 | ns/op |  0.12422360 | 0.99502488 |
| SlotMapBenchmark.embeddedQueryKey10Entries                  | N/A           | avgt |   5 |      0.664 |     0.006 |      0.673 |      0.008 | ns/op |  -1.3554217 |  1.1887073 |
| SlotMapBenchmark.hashInsert1Key                             | N/A           | avgt |   5 |      5.426 |     0.724 |      5.608 |      0.107 | ns/op |  -3.3542204 |  1.9079886 |
| SlotMapBenchmark.hashQueryKey100Entries                     | N/A           | avgt |   5 |      1.254 |     0.018 |      1.256 |      0.012 | ns/op | -0.15948963 | 0.95541401 |
| SlotMapBenchmark.hashQueryKey10Entries                      | N/A           | avgt |   5 |      1.131 |     0.011 |      1.172 |      0.017 | ns/op |  -3.6251105 |  1.4505119 |
| StartupBenchmark.startUpRhino                               | N/A           | avgt |   5 |     14.145 |     0.254 |     13.478 |      0.878 | us/op |   4.7154472 |  6.5143196 |
| SunSpiderBenchmark.AccessBinaryTreesState.accessBinaryTrees | false         | avgt |   5 |   4215.743 |    79.865 |   3792.279 |    261.463 | us/op |   10.044825 |  6.8946140 |
…