@gbrail 2024-12-30
This and other SlotMap changes look great, and I agree it's important to pay attention to this part of Rhino that's so performance-critical. However I don't want to rush these, and don't want to include them in 1.8.0, and will likely want to spend more time with them after the new year.