@aardvark179 2025-01-06
Made a minor edit to the PR description because I wrote the wrong size for embedded slot map, it was 32 bytes, not 24.