@gbrail 2025-01-09
Looks good so far! I appreciate the attention to detail and the documentation! I'll keep working through these.